+++ app/Filament/Resources/TicketRelationshipResource/Pages/CreateTicketRelationship.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\TicketRelationshipResource\Pages;
+
+use App\Filament\Resources\TicketRelationshipResource;
+use Filament\Resources\Pages\CreateRecord;
+
+class CreateTicketRelationship extends CreateRecord
+{
+    protected static string $resource = TicketRelationshipResource::class;
+}