+++ app/Filament/Resources/TicketRelationshipResource/Pages/EditTicketRelationship.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\TicketRelationshipResource\Pages;
+
+use App\Filament\Resources\TicketRelationshipResource;
+use Filament\Resources\Pages\EditRecord;
+
+class EditTicketRelationship extends EditRecord
+{
+    protected static string $resource = TicketRelationshipResource::class;
+}