+++ app/Filament/Resources/TicketRelationshipResource/Pages/ListTicketRelationships.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\TicketRelationshipResource\Pages;
+
+use App\Filament\Resources\TicketRelationshipResource;
+use Filament\Resources\Pages\ListRecords;
+
+class ListTicketRelationships extends ListRecords
+{
+    protected static string $resource = TicketRelationshipResource::class;
+}