+++ app/Http/Controllers/Api/TicketController.php
@@ -46,7 +46,13 @@ class TicketController extends Controller
 
         $this->authorizeForRequest($request, 'view', $ticket);
 
-        return TicketResource::make($ticket->load('assignee'));
+        return TicketResource::make($ticket->load([
+            'assignee',
+            'relationships.primaryTicket',
+            'relationships.relatedTicket',
+            'inverseRelationships.primaryTicket',
+            'inverseRelationships.relatedTicket',
+        ]));
     }
 
     public function update(UpdateTicketRequest $request, Ticket $ticket): TicketResource
@@ -55,7 +61,13 @@ class TicketController extends Controller
 
         $ticket = $this->service->update($ticket, $request->validated(), $request->user());
 
-        return TicketResource::make($ticket);
+        return TicketResource::make($ticket->load([
+            'assignee',
+            'relationships.primaryTicket',
+            'relationships.relatedTicket',
+            'inverseRelationships.primaryTicket',
+            'inverseRelationships.relatedTicket',
+        ]));
     }
 
     public function events(Request $request, Ticket $ticket): AnonymousResourceCollection