+++ app/Providers/AuthServiceProvider.php
@@ -7,10 +7,12 @@ use App\Models\KbArticle;
 use App\Models\Message;
 use App\Models\Ticket;
 use App\Models\TicketEvent;
+use App\Models\TicketRelationship;
 use App\Policies\ContactPolicy;
 use App\Policies\KbArticlePolicy;
 use App\Policies\MessagePolicy;
 use App\Policies\TicketEventPolicy;
+use App\Policies\TicketRelationshipPolicy;
 use App\Policies\TicketPolicy;
 use Illuminate\Foundation\Support\Providers\AuthServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\Gate;
@@ -20,6 +22,7 @@ class AuthServiceProvider extends ServiceProvider
     protected $policies = [
         Ticket::class => TicketPolicy::class,
         TicketEvent::class => TicketEventPolicy::class,
+        TicketRelationship::class => TicketRelationshipPolicy::class,
         Contact::class => ContactPolicy::class,
         KbArticle::class => KbArticlePolicy::class,
         Message::class => MessagePolicy::class,