+++ routes/api.php
@@ -3,6 +3,7 @@
 use App\Http\Controllers\Api\HealthcheckController;
 use App\Http\Controllers\Api\MessageController;
 use App\Http\Controllers\Api\TicketController;
+use App\Http\Controllers\Api\TicketRelationshipController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/v1/health', [HealthcheckController::class, 'show'])->name('api.health');
@@ -17,6 +18,11 @@ Route::middleware(['auth', 'tenant'])->prefix('v1')->name('api.')->group(functio
         Route::apiResource('tickets.messages', MessageController::class)->parameters([
             'tickets' => 'ticket',
             'messages' => 'message',
+        ]);
+
+        Route::apiResource('tickets.relationships', TicketRelationshipController::class)->parameters([
+            'tickets' => 'ticket',
+            'relationships' => 'relationship',
         ]);
     });
 });